Get program name of focused window if do_focus_id doesn’t provide that already (X11 only)
Currently works only in X11 sessions.

If a new enough version of ibus is used which supports focus id,

   do_focus_in_id(self, object_path: str, client: str)

has a “client” argument which indicates which input module is used
and sometimes also the program name in the active window.

I.e. one may get something like 'gtk3-im:firefox'.  But for the client
values 'xim', 'gnome-shell', 'Qt', 'QIBusInputContext', the program
name is missing.

If the client value does not contain a ':' it lacks the program name.
and this commit tries to add the program name in these cases

As this uses only the “xprop” binary at the moment, it currently works
only in Xorg sessions, not in Wayland sessions.

I hope I can find a way to make this work in Wayland sessions as well.