Make it possible to use custom labels for different dictionary matches
Often it is useful to distinguish in which dictionary or dictionaries
a candidate has been found.

There is an option “☑️ Use flags for dictionary suggestions” which can
help with that. But sometimes flags are not good enough to distinguish
dictionaries, see:

https://github.com/mike-fabian/ibus-typing-booster/issues/358

Even though in such a case the dictionary names are automatically aded
to the flags to show the difference, this is quite long and looks ugly.

This commit makes it possible to exactly specify which identifier should
be used to indicate matches in dictionaries. For example in the entry field
for

“☑️ Use label for dictionary suggestions”, instead of using a simple string
like the default “📖” one can now also specify a Python dictionary like this:

    {'*': '📖', 'fy_??': '🛟', 'de_DE': '🇩🇪', 'en_US': '🇺🇸'}

which means use “🛟” for matches in dictionaries matching the “fy_??”
pattern (for example the “fy_DE” and the “fy_NL” dictionaries), “🇩🇪”
for matches in the “de_DE” dictionary, “🇺🇸” for matches in the “en_US”
dictionary, and “📖” as a fallback for matches in other dictionaries.

If wildcards like “*” or “?” are used, only one match is used, *not*
all matches. I.e. even though “*” also matches “fy_DE”, only the most
specific match “fy_??” is used.

“most specific match” means the first matching key in the reverse sorted
list of keys, for example:

>>> sorted({'*', 'fy_*', 'fy_??', '?y_*', 'fy*', '??y_*', 'fy_DE', '*y_*'}, reverse=True)
['fy_DE', 'fy_??', 'fy_*', 'fy*', '?y_*', '??y_*', '*y_*', '*']

Here the “most specific match” for “fy_DE” is “fy_DE” and the “most
specific match” for “fy_NL” is “fy_??”.