Allow predictions when 0 characters have been typed
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/268

After a commit followed by white space, either because something like
`space` was typed to commit or by because a command like
`_command_commit_candidate_1_plus_space()` was executed because the
keybinding for such a command was typed, the preedit is empty and it
might be useful to suggest the next word already. Until now this was
not possible, a mininum of one new character typed was necessary to
enable a suggestion. The option

    Minimum number of chars for completion [1][-+]

could only be set to values from 1 to 9, not to 0.

Now this option can be set to 0. If this option has the value 0,
suggesting the next word is attempted afer a commit with white space
at the end.

As showing low quality suggestions would be useless and only annoying,
suggestions are only shown if a context of two previous words is available
and the user database contains suggestions following that context learned
from previous typing.

No enter return