emoji-picker: Put the first match in a search always into the clipboards automatically
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/402

Makes keyboard only search possible. As the first match is put always
automatially into the clipboard, there is no need anymore to click the
first match with the mouse.

Clicking with the mouse is still necessary if one wants to select any
of the matches following the first one.

But if one absolutely wants to avoid using the mouse, one could look
at the matches and refine the typed search string to make the desired
emoji the first match.