emoji-picker: If a search string contains non-whitespace characters, strip whitespace
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/404

If the string contains only whitespace, do not strip because just typing
space into the search field should match space like characters.

But if there are non-whitespace characters, stripping is better
because otherwise 'black cat ' would not be an exact match for 'black cat'.
That would create the weird effect that typing 'black cat' would
show 🐈‍⬛ as the first match but after adding a space to the search
string it would not be the first match anymore.