Save space when displaying compose completion in horizontal lookup tables
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/407

When the orientation of the lookup table is not set in the
ibus-typing-booster settings explicitly to either “Vertical” (the
default setting in ibus-typing-booster) or “Horizontal” but to “System
default”, then I have no way to find out whether the the lookup table
is actually horizontal or vertical. I can see only that the
orientation is “System default” but not whether that is actually
horizontal or vertical. Therefore, as “System default” could be
horizontal, I save space when showing compose completions in that case
as well. Only when I am sure that the lookup table is really
“Vertical”, I show the verbose information about the compose
completions.