Try not to fail if the database contains invalid UTF-8 or is otherwise malformed
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/409

I am not sure how this can happen, but using

        database_connection.text_factory = (
            lambda x: x.decode(errors='replace'))

will probably avoid an exception when the database contains invalid UTF-8.

To check for other problems than invalid UTF-8 I now try to read all
records from the database with `SELECT * FROM PHRASES;` during program
start.  If that produces an error, the database is probably broken. In
that case a backup copy of the broken database is made and a new empty
database is created.