Stop calling self.set_wmclass('ibus-setup-typing-booster', 'Typing Booster Preferences')
https://tronche.com/gui/x/icccm/sec-4.html#WM_CLASS

gnome-shell seemed to use the first argument of set_wmclass()
to find the .desktop file.

If the .desktop file is found, the name shown by gnome-shell in the
top bar comes from that .desktop file and the icon to show is also
read from that .desktop file. If the .desktop file cannot be found,
the second argument of set_wmclass() is shown by gnome-shell in the
top bar.

It only works like this when gnome-shell runs under Xorg
though, under Wayland things are different.

But https://docs.gtk.org/gtk3/method.Window.set_wmclass.html
says:

Don’t use this function. It sets the X Window System “class”
and “name” hints for a window. According to the ICCCM, you
should always set these to the same value for all windows in
an application, and GTK+ sets them to that value by default,
so calling this function is sort of pointless.

Deprecated since:       3.22
Please do not use it in newly written code.

Nevertheless calling set_wmclass() still used to be necessary
so I kept calling it inspite of the deprecation warning
on stdout.

But apparently this is not needed anymore since at least
Fedora 34/RHEL9, now the icon and the correct text appear in
the Gnome panel without calling set_wmclass().