Catch exception when user database is locked during shutdown
Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=2151448

do_destroy() in factory.py is called twice in quick succession during shutdown.
do_destroy() then calls tabsqlitedb.sync_usrdb() which does

   self.database.commit()
   self.database.execute('PRAGMA wal_checkpoint;')

Apparently this can cause a `sqlite3.OperationalError: database table is locked`.

But even if this commit and the checkpoint fail completely twice
there, this should not be a serious problem. And one of the calls
probably works anyway.

Therefore, I think it is better not to catch the exception and just
print a message into the log file then causing a traceback.