Do not save changes done by autosettings to gsettings
When the autosettings changes are written to gsettings, one can see
how the changes happen by keeping the setup tool open and looking at
the options.  That is nice, especially for debugging this feature.

*But* it seems to happen far too often that the changes done by
autosettings fail to properly be reverted to the standard settings on
“ibus restart” or when ibus is killed.

This is bad because then the autosettings intended only for a special
window become the new standard settings.

So better not write the autosettings changes to gsettings.