Autoselect extremely likely candidates: Allow only adding of accents, never removing accents
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/420

When the typed string contains no accents candidate just adds some
accents to the typed string, this can be useful in saving some
keystrokes.

In many cases, typing accents needs extra key strokes because dead
keys or compose are involved or AltGr has to be used or some prefix or
postfix input method has to be used.

So if one knows that a candidate which just adds the missing accents
will most likely be shown and probably even autoselected, one can
sometimes save some keystrokes by not bothering to type the accents at
all and just relying on Typing Booster adding the accents.

Sometimes this can also be helpful when one is using a keyboard layout
which cannot type the correct accents.

But the reverse seems almost never useful: When one has already
bothered to type accents into the typed string and the first candidate
just removes these accents, this is more often wrong than helpful.
The typed accents could of course be a spelling error and the first
candidate might fix that, but this seems rather unlikely.  It seems
more reasonable to assume that when the user already bothered to type
accents, then this more likely what the user wants than a candidate
removing these accents again.

Therefore, when the option

   Automatially select best candidate [Yes, but only when extremely likely]

is used, best candidates which just remove accents should not be considered.