Update the preedit to empty right after deleting surrounding text when reopening a preedit.
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/474

This resolves another part of https://github.com/mike-fabian/ibus-typing-booster/issues/474,
the problem seen in the screenshot in this comment:

https://github.com/mike-fabian/ibus-typing-booster/issues/474#issuecomment-1861837176

It seems to be a very rare problem, I could only reproduce it in the "Stacks" Editor ised
im https://meta.stackexchange.com/questions/ask .

There the problem occured that delete_surrounding_text() somehow
failed to remove the deleted text from the buffer, in the next call to
get_surrounding_text() the text seemed to be gone but in the editor
buffer it was still there.  But if after delete_surrounding_text() a non empty string is committed or
the show_preedit_text() is called or

update_preedit_text_with_mode(IBus.Text.new_from_string(''), 0, True,
                              IBus.PreeditFocusMode.COMMIT)

is called, then the deletion of the surrounding text works.

Unnecessary extra calls updating the preedit may trigger calls to
do_reset(), but this theoretically unnecessary update of the preedit
unfortunately seems to be the only way to fix the problem here.

If extra calls to do_reset() are caused by this, this should cause no problem
anymore  because

    commit f0a00abe0758a6d6e8b0fd38cf6a5ca7278f7bdf
    Author: Mike FABIAN <mfabian@redhat.com>
    Date:   Sat Dec 23 22:18:36 2023 +0100

        Improve do_reset()

improved do_reset() so that it does no harm when called with an empty preedit.