Add new combobox option “Record mode”
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/470

Chooses which words to record (words are called “phrases” in the source code for historic reasons):

    0: Everything
    1: Correctly spelled or previously recorded
    2: Correctly spelled
    3: Nothing

I left the existing option

    ☑️ Off the record mode

there for the moment. It works as it always did. Setting the new
“Record mode” option to 3 and checking “Off the record mode” does the
same thing so this seems a bit redundant.  But the advantage is that
one can quickly toggle recording to off using “Off the record mode”
with a keybinding and when toggling it on again return to the
operation of the selected “Record mode” in the combobox option.

I could maybe remove the option “3: Nothing” from the combobox to remove redundancy…

As I added one more row to the options tab of the setup tool, I saved one
row by putting the two buttons

[ Learn from text file ] [ Delete learned data ]

into the same row.