Introduce a delay time before candidates are updated and displayed
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/530

Until now I used

GLib.idle_add(self._update_candidates_and_lookup_table_and_aux)

to update the candidates and display a lookup table. This is to avoid
updating the candidates after every key press which might interfere
with fast typing. When keys are pressed in very fast succession, the
update of the candidates is delayed until the main loop is idle,
i.e. until there is at least a tiny pause between the key presses.

This used to be good enough, but maybe because computers have become
faster, that idle state in the main loop seems to be reached more
often now, on my machine it is very hard to type fast enough to avoid
reaching the idle loop between key presses. Also the calculation of
the candidate lists seems to be much faster than it used to be.

Therefore, I am introducing a delay time to update the candidates and
show a lookup table only if the main loop is idle for that delay time.

I think 200 milliseconds might be a reasonable default.

The new delay time is an option which can be set in the graphical
setup tool (in the appearances tab) or via gsettings:

```
$ gsettings set org.freedesktop.ibus.engine.typing-booster candidatesdelaymilliseconds 200
```

If that delay time is rather big, this could cause a delayed update of
the preedit as well because I tried to avoid duplicate updates of the
preedit and as _update_lookup_table() already updates the preedit,
updating it in _update_ui() is not strictly necessary if an update of
the candidates and the lookup table is already requested when the idle
loop is reached. But if the delay is long, waiting that long to update
the preedit is not OK anymore, so I have to do update the predit in
_update_ui always. This should be OK because I improved do_reset() in
the meantime to cause less problems even if an unwanted call to
do_reset() happens.

commit 2a54baeb0f083b4ba1058fbc7210d8878490322c
Author: Mike FABIAN <mfabian@redhat.com>
Date:   Mon Dec 18 16:57:12 2023 +0100

    Avoid more duplicate calls of _update_preedit()

    Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/474

    Don’t call _update_preedit() at the beginning of _update_ui(), only
    when returning without calling
    _update_candidates_and_lookup_table_and_aux().  _update_lookup_table()
    will call _update_preedit() (it sometimes has to because the inline
    completion option so it does it always, even when inline completion is
    not used). Therefore, if _update_preedit() is called at the beginning
    of _update_ui() and then _update_candidates_and_lookup_table_and_aux()
    then _update_preedit() is called twice in quick succession and this
    can trigger and unwanted call to do_reset() in some circumstances.