Get more information about the Pango rendering support for an emoji and use fallback in emoji-picker only when necessary
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/508

Using this it is possible to figure out whether a single emoji (which
might be a sequence of characters or a single code point emoji) or any
other Unicode character is supported by a certain font..

When it is known that an emoji is supported by the requested font,
Pango’s fallback can be disabled to avoid getting unneeded fallbacks.

Without that, rendering the following markup

<span font="Twemoji 48" fallback="false">🤯</span> Twemoji, fallback=false>

with Pango will fallback to using “Noto Color Emoji” even though the “Twemoji” font
supports 🤯 U+1F92F just fine.

This problem didn’t occur many years ago, but it happens with recent
versions of Pango and fontconfig (Happens for example on Fedora 40).

Also make it work for the latest version of the “Twitter Color Emoji”
and add a comment about that font.

Related Pango issues:

https://gitlab.gnome.org/GNOME/pango/-/issues/289
https://gitlab.gnome.org/GNOME/pango/-/issues/298