In do_reset() skip clearing input only when the surrounding text **up to the cursor** ends with the current preedit
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/529

Previously it was checked whether the complete surrounding test ended
with the current preedit still stored in self._current_preedit_text. If this is the case,
the input can be cleared because it has apparently been correctly committed.

This worked only correctly, if the current preedit happened to be at the end of the line.

But WhatsApp always seems to add a U+200B ZERO WIDTH SPACE after the
current cursor position. Therefore, when using WhatsApp in firefox,
the preedit was never really at the end of the line. When a preedit
which appeared to be at the end of the line was committed by clicking
with the mouse somewhere else in the line, the line actually ended
with the committed preedit followed by a U+200B ZERO WIDTH SPACE.

But the same problem can also be reproduced elsewhere in firefox and
also in gedit if the preedit is not at the end of the line.

I think the solution is to compare the contents of
self._current_preedit_text only with the surrounding text up to the
cursor positon.

Also, apply this workaround only when the im client uses gtk3-im, for
Wayland surrounding text works differently and has different contents
in this situation. The original problem the workaround is solving
occured on Gtk3, for other im clients that workaround is probably
useless anyway.