Improve “class KeyEvent”: Prepend 'S-' to msymbol when Shift is pressed and is is a space or not printable
Before I added 'S-' only for the special cases when msymbol was
'C-Return' or ' '.  Because 'S-C-Return' and 'S- ' were used in some
input methods in m17n-db.  But actually 'S-Left' and 'S-Right' are
also used already. And others like 'S-Up' or 'S-Down' could surely be
used. So the code handled only two special cases and was not generic enough.

But for characters which are printable and **not** whitespace, the
'S-' should **not** be added!  For example Ü (keysym is 'Udiaeresis')
is typed by pressing Shift+ü. And the Shift has then been “absorbed”
in making the characters uppercase, prepending 'S-' is then wrong.

See also https://github.com/ibus/ibus-m17n/issues/90 for the same problem in ibus-m17n.