Improve detection of Wayland session
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/561

There are situations where XDG_SESSION_TYPE=wayland is found in the ibus environment
even when an Xorg session is actually running. For example when this bug is present:

https://bugzilla.redhat.com/show_bug.cgi?id=2324342

It is better to test also whether WAYLAND_DISPLAY is set.

Testing for both WAYLAND_DISPLAY **and** XDG_SESSION_TYPE=wayland
should make the Wayland detection more reliable.