Improve emoji_font_fallback_needed() **always** ignore variation selectors when checking whether fallback is needed
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/565

Variation selectors seem to be always irrelevant for font fallback, if
they are present in an emoji sequence, pango may render it with more
than one glyph even when the sequences is fully supported. The
presence of the variation selector seems to cause the use of an
additional, apparently invisible glyph.