Hide “emoji prediction mode” and “off the record mode” property menus by default for the restricted engines emulating ibus-m17n
These two property menus are useless for a restricted engine which only
emulates ibus-m17n, there is no prediction so choosing “emoji
prediction mode” is useless and “off the record mode” is switched on
when ibus-m17n is emulated so this is also useless.

If the restrictions are lifted in the setup tool of a restricted engine,
the options become useful again.

So I making hiding the menus for these options depend on a combination
of option settings. An engine is surely restricted if all of the following
options are set:

- The option `tabenable` is set
- Only one input method is set
- No dictionary is set
- The command `enable_lookup` is not bound to any key