Prefer `return False` over `forward_key_event()` if the client is SDL2_Appliaction
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/580

Well, it resolves the problem only partly unfortunately. Because when
ibus-typing-booster is used for the very first time after a new
desktop session is started or for the very first time after ibus is
restarted, the client id ibus-typing-booster receives from ibus in
do_focus_in_id() is unfortunately empty.

So if an SDL2 application like supertuxkart is the very first
application used after a restart, the client id is empty and it is not
detected that it is an SDL2 application. But if ibus-typing-booster
has been already used in other windows before an SDL2 application like supertuxkart
is started, then the client id is `SDL2_Application` and ibus-typing-booster can
then prefer `return False` over `forward_key_event()`.