Use forward_key_event() in _return_false() only when fixing up wrong key codes
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/582

Related: https://github.com/mike-fabian/ibus-typing-booster/issues/580

If OSK produces key events with key codes 0, fixing the key codes and then
forwarding the fixed keys with forward_key_event() is the only way to make it work.
So until this is fixed in OSK, forward_key_event() has to be used there.

But the other cases where forward_key_event() had been necessary are
mostly not necessary anymore. For example when using XIM there was the
problem that when a commit was triggered by a space and then the space
send to the application using `return False`, the space arrived in the
applications before the commit and one got “ word” instead of the
correct “word ”. Using forward_key_event() to send the space was one
way of avoiding that problem, but for quite a while such spaces are
inserted by ibus-typing-booster into the application by committing a
space, which also avoids the problem. Using commits instead of sending
a key event seems to be the most reliable way to prevent the receiving
application to mess up the order of things. I think it has improved
things a lot especially on Wayland.