Add a gsettings option “prefercommit”
Related: https://github.com/mike-fabian/ibus-typing-booster/issues/580

This is to choose whether commits are preferred over passing through key events
in the _return_false() function.

Instead of passing a key event through, it is often possible to do a
commit.  And this is usually the best way to do it. It avoids several
problems like wrong order of events or wrong keycodes.  But there are
some cases where this may cause problems, especially in games which
expect real key events.

Therefore it can be useful to have an option available to switch off
preferring commits. If such an option is available, it can be switched
using autosettings.