Fix a few texts in the schema file to make them the same as in the setup tool
Related: https://github.com/mike-fabian/ibus-typing-booster/pull/604

This pull request originally wanted to make the schema file translatable
as well. That turned out to be useless at the moment as no tools
seem to be able to display texts from schema files when dynamic paths
are used. Typing Booster has to use dynamic paths as it offers many engines
which share the same schema file but can have different settings.

At the moment dconf-editor apparently cannot display texts from a
schema file at all if the schema file has no static path. I have no
idea whether it is possible fix this in dconf and as long as there are
no tools reading and displaying the texts from the schema file,
translating it seems useless.

But if this ever changes and we want to translate the schema file, the
texts in the schema file should not gratuitously differ from the texts
in the setup tool.  If the texts are identical as far as possible,
this reduces the work for the translators.

I had already started making the texts in the schema file identical to
those in the setup tool before I realized that translations is useless
at the moment.

Committing these changes anyway cannot hurt, it might be helpful if
the situation changes and translating the schema file becomes useful.