Add a new gsettings option 'wordpredictions' to simplify and improve the behaviour of engines emulating ibus-m17n
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/629

Instead of using the weird combination of options

    bool(self._tab_enable
         and self._off_the_record
         and len(self._current_imes) == 1
         and self._dictionary_names == ['None']
         and not self._keybindings['enable_lookup'])

to figure out whether it is one of the engines emulating the behaviour
of ibus-m17n, introduce a new gsettings option 'wordpredictions'. This
makes the code more logical, now more reasonable checks whether
prediction candidates shold be displayed and whether it makes sense to
try an early commit (like ibus-m17n does) are possible:

    def _show_prediction_candidates(self) -> bool:
        '''Checks whether prediction candidate lists are wanted

        If neither emoji predictions are possible nor word predictions
        are possible, calculations to produce the “normal” Typing Booster
        lookup table can be skipped completely.

        But m17n candidates lookup tables and Compose lookup tables
        are still possible then.
        '''
        return bool(self._emoji_predictions
                    or self._emoji_trigger_characters != ''
                    or self._word_predictions)

    def _try_early_commit(self) -> bool:
        '''
        Whether it should be checked if parts of the
        transliteration can be committed already
        '''
        return bool(not self.is_empty()
                    and len(self._current_imes) == 1
                    and not self._typed_compose_sequence
                    and not self._m17n_trans_parts.candidates
                    and not self._word_predictions
                    and not self._emoji_predictions
                    and not self._typed_string[0] in self._emoji_trigger_characters)

That is not only more logical and readable but it also makes it
possible to always have the same defaults for 'tabenable' and the
keybinding for 'enable_lookup', for the engines emulating ibus-m17n as
well as for the “normal” Typing Booster:

tabenable: false
'enable_lookup': ['Tab', 'ISO_Left_Tab']

That makes it possible for engines emulating ibus-m17n like

    tb:t:rfc1345
    tb:vi:telex
    ...

to show m17n candidate lists for fallback input methods like t-lsymbol
by default. ibus-m17n does the same, i.e. this makes the emulation of
ibus-m17n better.

It also makes it possible to enable m17n input methods offering multiple
candidates as “restricted” engines, i.e. it is now possible to enable
these engines as “restricted” engines emulating the respective ibus-m17n engines:

    tb:ja:anthy
    tb:t:lsymbol
    tb:ml:swanalekha
    tb:vi:han
    tb:vi:nomtelex
    tb:vi:nomvni
    tb:zh:cangjie
    tb:zh:py
    tb:zh:py-b5
    tb:zh:py-gb
    tb:zh:quick
    tb:zh:tonepy
    tb:zh:tonepy-gb
    tb:zh:tonepy-b5
    tb:zh:zhuyin

That makes the emulation of ibus-m17n more complete. Note that a few
of these get negative ranks though:

    (r'tb:ja:anthy', -1), # ja:anthy is superseded by ibus-anthy
    (r'tb:zh:py', -1), # zh:py is superseded by ibus-pinyin etc
    (r'tb:zh:py-gb', -1), # zh:py-gb is superseded by ibus-pinyin etc
    (r'tb:zh:py-b5', -1), # zh:py-b5 is superseded by ibus-pinyin etc
    (r'tb:ko:han2', -1), # ko:han2 is superseded by ibus-hangul
    (r'tb:ko:romaja', -1), # ko:romaja is superseded by ibus-hangul

and therefore they don’t show up in `ibus list-engine`. This is the same
as in ibus-m17n, all the ranks are the same as in ibus-m17n.

Another thing which is made possible by this change is doing something like this:

- use t-rfc1345 input method with the following settings:
- ☐ Word predictions
- ☐ Unicode symbols and emoji predictions
- Emoji trigger characters '_:'

This makes it possible using t-rfc1345 while almost never showing
candidate lists except for when a fallback input method like t-lsymbol
activates or when a word is started with an emoji trigger
character. If a word is started with an emoji trigger character, early
commits don’t happen, the word is kept in preedit and emojis are
searched.