Use itb_util.normalize_nfc_and_composition_exclusions(x) *everywhere* instead of unicodedata.normalize('NFC', x)
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/633

I introduced itb_util.normalize_nfc_and_composition_exclusions(x)
to fix:

https://github.com/mike-fabian/ibus-typing-booster/issues/531

but I used it only for committed strings to fix for performance reasons.

I thought it should be enough to do this for committed strings because the
commits is what the user gets into the application in the end.
Just normalizing to 'NFC' should be enough for the preedit, the caret in the preedit, the candidates, and the lookup table, I thought.
Because these three Bengali characters

ড় 09DC # BENGALI LETTER RRA
ঢ় 09DD # BENGALI LETTER RHA
য় 09DF # BENGALI LETTER YYA

which are listed in
https://www.unicode.org/Public/16.0.0/ucd/CompositionExclusions.txt as
composition exclusions, visually display the same way in the preedit
and in the lookup tables no matter whether they are precomposed or
not. So I thought I could save some time by doing the extra work
to force composition of the composition exclusions only for the commits.

I still think that should be enough, but apparently WhatsApp and
Facebook in google-chrome do care for the exact contents of the
preedit: if the preedit contains a string where a non-precomposed of
one of these Bengali characters is the 3rd syllable and then that
string is committed in the precomposed form, then it creates a mess in
WhatsApp and Facebook in google-chrome.

Apparently this can be avoided by having the exact same code points
everywhere in the commit, the preedit, the caret, the candidates, the
lookup table.

I don’t think this has any disadvantages except a small loss in
performance.