Ignore dictionary matches in the lookup table for candidates which have comments
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/655

If a candidate in a lookup table has a comment, then it must be

- an emoji
- a single “unusual” charactor or symbol
- a related word found by itb_nltk.py

In all these cases, it is not interesting to apply labels to the
lookup table for such candidates to show whether a spellcheck against
some language specific dictionaries returns True or False.

On some systems, something like 💩 might pass a spellcheck, for example on Debian:

    mfabian@debian-testing:~$ python3
    Python 3.13.2 (main, Feb  5 2025, 01:23:35) [GCC 14.2.0] on linux
    Type "help", "copyright", "credits" or "license" for more information.
    >>> import enchant
    >>> d = enchant.Dict('en_US')
    >>> d.check('💩')
    True
    >>>

But that doesn’t  mean it is useful to append a 🇺🇸 label to a 💩 to
indicate that 💩 is a valid word in the en_US dictionary.