Display a middle dot in the preedit for keys in a compose sequence which use unassigned Unicode code points.
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/674

I treat these unassigned code points the same way as Multi_key
now:

For Multi_key, a middle dot is displayed.  If the preedit
representation of a longer sequence would with a middle dot, the first
middle dot in that sequence is suppressed all other middle dots but
the first are shown.

These keys with unassigned Unicode code points are now treated the same way.