Some optimization and code cleanup of the emoji search code
I removed the old `import SequenceMatcher` code, it has been disabled
for years already because it was slow and not so useful.  Removing
that I could refactor the code a bit and get rid of `self._set_seq1()`
and `self._set_seq2()` which made it easier to use
@functools.lru_cache instead of manual caching. Overall this gave a
performance improvement of about 20-30% while keeping the all match
scores exactly the same, i.e. all test cases still work with the same
expected results.

I also tried David Mandelberg’s suggestion to match emoji
spellchecking first and only if there is no match at all, try again
with spellchecking.

But actually this makes the performance somewhat worse, so I disabled
it again.

It makes the performance worse because when a misspelled word is
searched for, the search may loop over all the emoji data twice,
first matching against the mispelled word which might return no
results and then again searching for matches against all the all the
spellchecking suggestions.

Of course it may shortcut if the original search input already returns
a result, in that case it is somewhat faster because the spellchecking
suggestions are not added to the words which are used for matching.

So it depends on the exact test input whether this is an performance
improvement or makes the performance worse, it makes the performance
worse only if there is a significant amount of input which does not
match already without matching the spellchecking suggestions as well.

But for the test input I currently have in the doctests of
itb_emoji.EmojiMatcher.candidates() this makes the performance about
15% worse.

I am not sure whether that test input is representative of typical
input when typing, it is hard to say whether this improves performance
or makes it worse on typical typing input. Anyway the difference is
not huge either way.