Show number of matched emoji always in the subtitle in emoji-picker and increase matchlimit to 100_000 for searches
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/689

The number of emoji shown in a list of matches disappeared after selecting an emoji.
I.e. when the `people` category was selected, the header bar first showed

      people
       606

But when then an emoji was selected with the mouse, the header bar changed to display
the emoji with its names in the title and the list of  keywords in the subtitle. I.e. something like this:

    🤩 smiling face with heart-shaped eyes, smiling face with heart-eyes, heart eyes
    Keywords: eye, face, love, smile, uc6, 143, bae, feels, heart-eyes, hearts, ily, kisses, romance, romantic, xoxo

(Elided at the end because it is usually too long)

Then the number of emoji in the list displayed had disappeared. When selecting a category like “people” this is not a big problem
because one has seen that number before selecting the emoji and one can still see the number in the sidebar on the left.

But when the list of emoji displayed is the result of a search string,
the first result is automatically selected.  This is on purpose to
have the first result automatically added to the clipboard so one can
paste the first result immediately somewhere else without needing an
extra mouse click to select it.

But because of that automatic selection of the first result on
searches, the number of matches for the search displayed in the
subtitle of the header bar disappeared immediately.

I want to see the number of matches of a search.

So I changed it to display the number of emoji in a list of matches
**always** (in parentheses) at the start of the subtitle now.

I.e. the header bar may look like this now:

      people
      (606)

or

    🤩 smiling face with heart-shaped eyes, smiling face with heart-eyes, heart eyes
    (606) Keywords: eye, face, love, smile, uc6, 143, bae, feels, heart-eyes, hearts, ily, kisses, romance, romantic, xoxo

or when searching for “horse”

    🐴 horse face, horse
    (14) Keywords: face, horse, uc6, animal, dressage, equine, farm, horses

On top of that I changed the match_limit for emoji searches in
emoji-picker from 1000 to 100_000.  This should be enough to never cut
any list of matches short in emoji picker (The search string
“egyptian” gives me 5105 matches, that is the longest I could find).